+++ pkg/kubernetes/delete_resources.go
@@ -6,7 +6,6 @@ import (
 	"log"
 
 	apiv1 "k8s.io/api/core/v1"
-
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
